refactor(profiles): split mutations file + consolidate duplicate role mutations
Split the ~745-line baseapp_profiles/graphql/mutations.py into a mutations/
package (profiles.py, roles.py, invitations.py) with __init__ re-exporting the
public surface + the ProfilesMutations aggregator, so the dotted import path
(baseapp_profiles.graphql.mutations.ProfilesMutations / ProfileCreateSerializer)
stays unchanged.

Remove two duplicated mutations:
- RoleUpdate (profileRoleUpdate) duplicated ProfileUserRoleUpdate — kept the
  consistently-named ProfileUserRoleUpdate and folded in RoleUpdate's robustness
  (role_type required + validated, not-found code).
- ProfileRemoveMember (profileRemoveMember) duplicated ProfileUserRoleDelete and
  was buggy (.get() with no try/except) — kept ProfileUserRoleDelete.

Repoint the role-update tests to profileUserRoleUpdate; drop the duplicate
profileRemoveMember tests (profileUserRoleDelete is already covered in
test_graphql_mutations_delete.py). All 131 baseapp_profiles tests pass.
