+++ baseapp_profiles/tests/test_graphql_mutations_update.py
@@ -36,8 +36,8 @@ PROFILE_UPDATE_GRAPHQL = """
 """
 
 PROFILE_ROLE_UPDATE_GRAPHQL = """
-mutation ProfileRoleUpdateMutation($input: RoleUpdateInput!) {
-    profileRoleUpdate(input: $input) {
+mutation ProfileUserRoleUpdateMutation($input: ProfileUserRoleUpdateInput!) {
+    profileUserRoleUpdate(input: $input) {
         profileUserRole {
             id
             role
@@ -47,14 +47,6 @@ mutation ProfileRoleUpdateMutation($input: RoleUpdateInput!) {
             field
             messages
         }
-    }
-}
-"""
-
-PROFILE_MEMBER_REMOVE_GRAPHQL = """
-mutation ProfileRemoveMemberMutation($input: ProfileRemoveMemberInput!) {
-    profileRemoveMember(input: $input) {
-        deletedId
     }
 }
 """
@@ -253,7 +245,7 @@ def test_user_profile_owner_can_update_role(django_user_client, graphql_user_cli
     )
     content = response.json()
 
-    assert content["data"]["profileRoleUpdate"]["profileUserRole"]["role"] == "ADMIN"
+    assert content["data"]["profileUserRoleUpdate"]["profileUserRole"]["role"] == "ADMIN"
     profile.refresh_from_db()
 
 
@@ -280,7 +272,7 @@ def test_user_with_permission_can_update_role(django_user_client, graphql_user_c
     )
     content = response.json()
 
-    assert content["data"]["profileRoleUpdate"]["profileUserRole"]["role"] == "ADMIN"
+    assert content["data"]["profileUserRoleUpdate"]["profileUserRole"]["role"] == "ADMIN"
     profile.refresh_from_db()
 
 
@@ -308,76 +300,3 @@ def test_user_without_permission_cant_update_role(django_user_client, graphql_us
     content = response.json()
     assert content["errors"][0]["message"] == "You don't have permission to perform this action"
     profile.refresh_from_db()
-
-
-def test_user_profile_owner_can_remove_profile_member(django_user_client, graphql_user_client):
-    perm = Permission.objects.get(
-        content_type__app_label=ProfileUserRole._meta.app_label, codename="delete_profileuserrole"
-    )
-
-    user = django_user_client.user
-    user_2 = UserFactory()
-
-    user.user_permissions.add(perm)
-    profile = ProfileFactory(owner=user)
-    profile_user_role = ProfileUserRoleFactory(
-        profile=profile, user=user_2, role=ProfileUserRole.ProfileRoles.MANAGER
-    )
-    profile_user_role_relay_id = profile_user_role.relay_id
-
-    response = graphql_user_client(
-        PROFILE_MEMBER_REMOVE_GRAPHQL,
-        variables={"input": {"userId": user_2.relay_id, "profileId": profile.relay_id}},
-    )
-    content = response.json()
-
-    assert content["data"]["profileRemoveMember"]["deletedId"] == profile_user_role_relay_id
-    assert not ProfileUserRole.objects.filter(id=profile_user_role.id).exists()
-
-
-def test_user_with_permission_can_remove_profile_member(django_user_client, graphql_user_client):
-    perm = Permission.objects.get(
-        content_type__app_label=ProfileUserRole._meta.app_label, codename="delete_profileuserrole"
-    )
-
-    user = django_user_client.user
-    user.user_permissions.add(perm)
-    user_2 = UserFactory()
-    user_3 = UserFactory()
-
-    profile = ProfileFactory(owner=user_2)
-    ProfileUserRoleFactory(profile=profile, user=user, role=ProfileUserRole.ProfileRoles.ADMIN)
-    profile_user_role = ProfileUserRoleFactory(
-        profile=profile, user=user_3, role=ProfileUserRole.ProfileRoles.MANAGER
-    )
-    profile_user_role_relay_id = profile_user_role.relay_id
-
-    response = graphql_user_client(
-        PROFILE_MEMBER_REMOVE_GRAPHQL,
-        variables={"input": {"userId": user_3.relay_id, "profileId": profile.relay_id}},
-    )
-    content = response.json()
-    assert content["data"]["profileRemoveMember"]["deletedId"] == profile_user_role_relay_id
-    assert not ProfileUserRole.objects.filter(id=profile_user_role.id).exists()
-
-
-def test_user_without_permission_cant_remove_profile_member(
-    django_user_client, graphql_user_client
-):
-    user = django_user_client.user
-    user_2 = UserFactory()
-    user_3 = UserFactory()
-
-    profile = ProfileFactory(owner=user_2)
-    ProfileUserRoleFactory(profile=profile, user=user, role=ProfileUserRole.ProfileRoles.MANAGER)
-    profile_user_role = ProfileUserRoleFactory(
-        profile=profile, user=user_3, role=ProfileUserRole.ProfileRoles.MANAGER
-    )
-
-    response = graphql_user_client(
-        PROFILE_MEMBER_REMOVE_GRAPHQL,
-        variables={"input": {"userId": user_3.relay_id, "profileId": profile.relay_id}},
-    )
-    content = response.json()
-    assert content["errors"][0]["extensions"]["code"] == "permission_required"
-    assert ProfileUserRole.objects.filter(id=profile_user_role.id).exists()
